+++ js/components/index.js
@@ -1,0 +1,9 @@
+import NumberSpinner from './NumberSpinner';
+import Player from './Player';
+import SpinnerButton from './SpinnerButton';
+
+export {
+    NumberSpinner,
+    Player,
+    SpinnerButton,
+};
